+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 * ⚠️ **BREAKING CHANGE:** If left unspecified, slash command `defaultMemberPermissions` now defaults to `"0"` i.e. admins only
 * ⚠️ **BREAKING CHANGE:** Renamed `Knub.stop()` to `Knub.destroy()`
   * Matching naming with djs's `Client.destroy()`, and more clearly shows that the Knub instance is unusable after this
+* ⚠️ **BREAKING CHANGE:** Removed `info` property from PluginBlueprint
+  * This property really only existed for Zeppelin. If you need to add it back, extend the interface.
 * Context menu commands are now supported
   * Helpers: `guildPluginMessageContextMenuCommand()`, `guildPluginUserContextMenuCommand()`
   * Context menu blueprints are listed in the `contextMenuCommands` property of a plugin